Allow longer debounce

+++ santroller-configurator/Configuration/Outputs/ControllerButton.cs
@@ -14,7 +14,7 @@ public class ControllerButton : OutputButton
 {
 
     public ControllerButton(ConfigViewModel model, Input input, Color ledOn, Color ledOff, byte[] ledIndices, byte[] ledIndicesPeripheral,
-        byte debounce, StandardButtonType type, bool childOfCombined) : base(model, input, ledOn, ledOff, ledIndices, ledIndicesPeripheral,
+        int debounce, StandardButtonType type, bool childOfCombined) : base(model, input, ledOn, ledOff, ledIndices, ledIndicesPeripheral,
         debounce, childOfCombined)
     {
         Type = type;
