+++ santroller-configurator/Configuration/Outputs/GuitarButton.cs
@@ -15,7 +15,7 @@ public class GuitarButton : OutputButton
 {
     public readonly InstrumentButtonType Type;
 
-    public GuitarButton(ConfigViewModel model, Input input, Color ledOn, Color ledOff, byte[] ledIndices, byte[] ledIndicesPeripheral, byte debounce,
+    public GuitarButton(ConfigViewModel model, Input input, Color ledOn, Color ledOff, byte[] ledIndices, byte[] ledIndicesPeripheral, int debounce,
         InstrumentButtonType type, bool childOfCombined) : base(model, input, ledOn, ledOff, ledIndices, ledIndicesPeripheral, debounce,
         childOfCombined)
     {
